+++ src/Field/FieldBufferOps.hpp
@@ -70,8 +70,8 @@ namespace ippl {
             size_t size = intersect.size();
             nsends      = size;
             if (buffer.size() < size) {
-                const int overalloc = ippl::Comm->getDefaultOverallocation();
-                Kokkos::realloc(buffer, size * overalloc);
+                const double overalloc = ippl::Comm->getDefaultOverallocation();
+                Kokkos::realloc(buffer, static_cast<size_t>(size * overalloc));
             }
 
             const int first0 = intersect[0].first() + nghost - ldom[0].first();
@@ -114,8 +114,8 @@ namespace ippl {
             size_t size = intersect.size();
             nsends      = size;
             if (buffer.size() < size) {
-                const int overalloc = ippl::Comm->getDefaultOverallocation();
-                Kokkos::realloc(buffer, size * overalloc);
+                const double overalloc = ippl::Comm->getDefaultOverallocation();
+                Kokkos::realloc(buffer, static_cast<size_t>(size * overalloc));
             }
 
             const int first0 = intersect[0].first() + nghost - ldom[0].first();
@@ -146,7 +146,7 @@ namespace ippl {
 
         // Unpack a linear buffer into a 3D view region, with per-axis conditional
         // reflection of the buffer index. Setting x / y / z to true reverses the
-        // buffer ordering along that axis as it is placed into the view — the
+        // buffer ordering along that axis as it is placed into the view -- the
         // primitive operation behind `mirrorField` and behind the Vico solver's
         // reflected-quadrant assembly.
         template <int tensorRank, typename Tb, typename Tf>
